py/requirements.txt
PySide6
wikipedia
sentence-transformers
ddgs
torch
gliner
nltk
faster-whisper
requests
huggingface-hub
fastapi
uvicorn
python-multipart
av
Pillow
posthog